Create readme.md
This folder contains Java programs that demonstrate the use of recursion to solve problems. The programs cover a variety of concepts, from simple recursive functions to more complex problem-solving approaches. The aim is to strengthen understanding of how recursion works and how it can simplify problems that involve repeated sub-tasks.